+++ packages/server/prod-server/src/libs/render/ssr.ts
@@ -1,5 +1,4 @@
 import path from 'path';
-import { IncomingMessage } from 'http';
 import {
   fs,
   mime,
@@ -45,7 +44,6 @@ export const render = async (
   const routeManifest = fs.existsSync(routesManifestUri)
     ? require(routesManifestUri)
     : undefined;
-  const body = await getRequestBody(ctx.req);
 
   const context: SSRServerContext = {
     request: {
@@ -56,7 +54,7 @@ export const render = async (
       query: ctx.query as Record<string, string>,
       url: ctx.href,
       headers: ctx.headers,
-      body,
+      body: ctx.req.body,
     },
     response: {
       setHeader: (key, value) => {
@@ -112,28 +110,3 @@ export const render = async (
     };
   }
 };
-
-/**
- * get body from request.
- * @return body -
- *  if req.method !== 'GET', it returns a string, otherwise it returns undefined
- */
-const getRequestBody = (req: IncomingMessage): Promise<string | undefined> =>
-  new Promise((resolve, reject) => {
-    if (req?.method && req.method.toLowerCase() !== 'get') {
-      let body = '';
-      req.on('data', chunk => {
-        body += chunk.toString();
-      });
-
-      req.on('end', () => {
-        resolve(body);
-      });
-
-      req.on('error', err => {
-        reject(err);
-      });
-    } else {
-      resolve(undefined);
-    }
-  });
