+++ packages/server/prod-server/src/server/modernServer.ts
@@ -46,6 +46,7 @@ import {
   noop,
   debug,
   isRedirect,
+  bodyParser,
 } from '../utils';
 import * as reader from '../libs/render/reader';
 import { createProxyHandler, BffProxyOptions } from '../libs/proxy';
@@ -503,7 +504,10 @@ export class ModernServer implements ModernServerInterface {
   /* —————————————————————— private function —————————————————————— */
   // handler route.json, include api / csr / ssr
   private async routeHandler(context: ModernServerContext) {
-    const { res } = context;
+    const { res, req } = context;
+
+    // parse request body from readable stream to assgin it to req
+    await bodyParser(req);
 
     // match routes in the route spec
     const matched = this.router.match(context.path);
